+++ SavingYourBacon/src/app/incomes/incomes.component.ts
@@ -1,14 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-
 import { Income } from '../data-objects/income';
-
 import { IncomeService } from '../services/income-service.service';
-
 import { IncomeRecord } from '../data-objects/IncomeRecord';
+import { MdDialog, MdDialogRef, MdDialogConfig, MdSnackBar } from '@angular/material';
+import { ConfirmDialogComponent } from '../common/confirm-dialog.component';
+import { ConfirmDialog as ConfirmDialogModel } from '../data-objects/confirmDialog';
 
-import {MdSnackBar} from '@angular/material';
-
-declare let _:any;
+declare let _: any;
 
 @Component({
   selector: 'incomes',
@@ -23,40 +21,52 @@ export class IncomesComponent implements OnInit {
   mode = 'Observable';
   userId = 1000;
 
-  constructor(private incomeService: IncomeService, public snackBar: MdSnackBar) { }
+  constructor(private incomeService: IncomeService, public snackBar: MdSnackBar, public dialog: MdDialog) { }
 
   ngOnInit() {
-      this.getIncome();
+    this.getIncome();
   }
 
-  getIncome(){
-      this.incomeService.getIncome(this.userId)
-                        .subscribe(data => this.income = data,
-                                    error =>  this.errorMessage = <any>error);
+  getIncome() {
+    this.incomeService.getIncome(this.userId)
+      .subscribe(data => this.income = data,
+      error => this.errorMessage = <any>error);
   }
 
-  handleIncomeEdit(event){
+  handleIncomeEdit(event) {
 
     let that = this;
 
     this.incomeService.updateIncome(event.data)
-                        .subscribe(data => {
-                                      that.snackBar.open('Income record updated successfully!', '', { duration: 1000 });
-                                    },
-                                    error =>  this.errorMessage = <any>error);
+      .subscribe(data => {
+        that.snackBar.open('Income record updated successfully!', '', { duration: 1000 });
+      },
+      error => this.errorMessage = <any>error);
   }
 
-  deleteIncome(income: IncomeRecord){
+  deleteIncome(income: IncomeRecord) {
     let that = this;
-    
-    this.incomeService.deleteIncome(income.IncomeId)
-                      .subscribe(data => { 
-                                            _.remove(this.income, function(currentObject) {
-                                                return currentObject.IncomeId === income.IncomeId;
-                                            });
-                                            that.snackBar.open('Income record deleted successfully!', '', { duration: 1000 }); 
-                                         },
-                                 error => this.errorMessage = <any>error);
+    let model: ConfirmDialogModel = { Title: "Are you sure?", Body: "" };
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, { data: model });
+
+    dialogRef.afterClosed()
+      .subscribe(data => {
+        if (data) {
+          this.incomeService.deleteIncome(income.IncomeId)
+            .subscribe(data => {
+              _.remove(this.income, function (currentObject) {
+                return currentObject.IncomeId === income.IncomeId;
+              });
+              that.snackBar.open('Income record deleted successfully!', '', { duration: 1000 });
+            },
+            error => this.errorMessage = <any>error);
+
+        } else {
+          return false;
+        }
+
+      }, error => this.errorMessage = <any>error
+      );
   }
 
 }
